test(agent): vitest suite + CI for the agent runner; fix relay error bug
Convert the hand-run tsx/node:assert scripts under services/agent/test/ into a vitest suite
in tests/unit/ (describe/it, node:assert kept). Add vitest + typescript devDeps, scripts
(test/typecheck/coverage), a node-env vitest.config (junit + v8 coverage), and broaden the
tsconfig include so typecheck covers tests + config. Add a run-services-node-unit-tests job to
12-check-unit-tests.yml, and a services/agent/AGENTS.md.

Testability seam: server.ts exports createAgentServer(run); cli.ts exports
runCli(raw, stream, io) with an injectable engine + output sink, so the HTTP and CLI paths are
tested with a fake engine (server.test.ts, cli.test.ts). src/entry.ts isEntrypoint keeps the
entrypoints inert on import. Add src/version.ts + a richer GET /health; version 0.0.0 -> 0.1.0.

Fix a real bug in tools/dispatch.ts: relayToolCall referenced an undefined callRef in its
error/timeout messages (would throw ReferenceError and mask the real failure); use toolName,
plus a focused relay regression test.

Result: 10 test files / 47 tests pass, tsc --noEmit clean. Two tests are deferred because their
deps live on sibling branches: skills.test.ts (needs engines/skills.ts from feat/agenta-on-rivet)
and wire-contract.test.ts (needs the shared Python golden fixtures). They land when those reach
this branch.
