+++ .gitignore
@@ -65,6 +65,8 @@ sdks/python/oss/tests/results/
 sdks/python/ee/tests/results/
 services/oss/tests/results/
 services/ee/tests/results/
+services/agent/test-results/
+services/agent/coverage/
 .*
 !**/.gitkeep
 !.github/
