+++ services/agent/src/cli.ts
@@ -1,28 +1,91 @@
 /**
  * WP-2 Pi wrapper CLI: the JSON transport for the Harness port.
  *
- * Reads one JSON `AgentRunRequest` from stdin, runs Pi once, and writes one JSON
- * `AgentRunResult` to stdout. stdout carries the result and nothing else; logs go
- * to stderr. This is the one-shot "json adapter" the design doc describes; a
- * long-lived RPC adapter can replace it later behind the same Python-side port.
+ * Reads one JSON `AgentRunRequest` from stdin, runs the agent once, and writes one JSON
+ * `AgentRunResult` to stdout. stdout carries the result and nothing else; logs go to stderr.
+ * With `--stream`, writes NDJSON instead: one `{kind:"event"}` line per event the moment it
+ * is built, then exactly one terminal `{kind:"result"}` line.
+ *
+ * `runCli(raw, stream, io)` is the testable seam: it takes the raw stdin string and an
+ * injectable engine runner + output sink, and returns the exit code. Tests pass a fake engine
+ * and a collecting `write`, so no stdin/stdout/process.exit mocking is needed; production
+ * defaults to the real engine and `process.stdout` (which keeps streaming live).
  */
 import type {
   AgentRunRequest,
   AgentRunResult,
   EmitEvent,
-  StreamRecord,
 } from "./protocol.ts";
 import { runPi } from "./engines/pi.ts";
 import { runRivet } from "./engines/rivet.ts";
+import { isEntrypoint } from "./entry.ts";
+
+/** Run one request through an engine. Tests inject a fake to avoid a live harness. */
+export type RunAgent = (
+  request: AgentRunRequest,
+  emit?: EmitEvent,
+) => Promise<AgentRunResult>;
 
 // Engine: `rivet` drives a harness over ACP via a rivet daemon; `pi` (default) is the
 // legacy in-process Pi path. The request's `backend` wins, then the AGENT_BACKEND env.
-function runAgent(
-  request: AgentRunRequest,
-  emit?: EmitEvent,
-): Promise<AgentRunResult> {
+const runAgent: RunAgent = (request, emit) => {
   const backend = (request.backend ?? process.env.AGENT_BACKEND ?? "pi").toLowerCase();
   return backend === "rivet" ? runRivet(request, emit) : runPi(request, emit);
+};
+
+function errorMessage(err: unknown): string {
+  return err instanceof Error ? err.stack ?? err.message : String(err);
+}
+
+export interface CliIO {
+  /** Engine runner; defaults to the real backend dispatch. */
+  run?: RunAgent;
+  /** Output sink; defaults to `process.stdout`. Called incrementally so streaming stays live. */
+  write?: (chunk: string) => void;
+}
+
+/**
+ * Run one request and return the process exit code (0 = ok, 1 = failure/invalid input).
+ * Output is delivered through `io.write` as it is produced.
+ */
+export async function runCli(
+  raw: string,
+  stream: boolean,
+  io: CliIO = {},
+): Promise<number> {
+  const run = io.run ?? runAgent;
+  const write = io.write ?? ((chunk: string) => void process.stdout.write(chunk));
+
+  let request: AgentRunRequest;
+  try {
+    request = raw.trim() ? (JSON.parse(raw) as AgentRunRequest) : {};
+  } catch (err) {
+    const failure: AgentRunResult = { ok: false, error: `Invalid JSON on stdin: ${String(err)}` };
+    write(stream ? JSON.stringify({ kind: "result", result: failure }) + "\n" : JSON.stringify(failure));
+    return 1;
+  }
+
+  if (!stream) {
+    try {
+      const result = await run(request);
+      write(JSON.stringify(result));
+      return result.ok ? 0 : 1;
+    } catch (err) {
+      write(JSON.stringify({ ok: false, error: errorMessage(err) }));
+      return 1;
+    }
+  }
+
+  const emit: EmitEvent = (event) => write(JSON.stringify({ kind: "event", event }) + "\n");
+  let result: AgentRunResult;
+  try {
+    result = await run(request, emit);
+  } catch (err) {
+    result = { ok: false, error: errorMessage(err) };
+  }
+  // Streaming delivered the events live, so don't echo them in the terminal record.
+  write(JSON.stringify({ kind: "result", result: { ...result, events: [] } }) + "\n");
+  return result.ok ? 0 : 1;
 }
 
 async function readStdin(): Promise<string> {
@@ -33,56 +96,14 @@ async function readStdin(): Promise<string> {
   return Buffer.concat(chunks).toString("utf8");
 }
 
-// One-shot mode: the whole result as a single JSON document (the `/invoke` contract).
-function emitResult(result: AgentRunResult): void {
-  process.stdout.write(JSON.stringify(result));
-}
-
-// Streaming mode (`--stream`): one NDJSON record per line — an `{kind:"event"}` line the
-// moment each event is built, then exactly one terminal `{kind:"result"}` line.
-function writeRecord(record: StreamRecord): void {
-  process.stdout.write(JSON.stringify(record) + "\n");
-}
-
 async function main(): Promise<void> {
   const stream = process.argv.includes("--stream");
   const raw = await readStdin();
-
-  let request: AgentRunRequest;
-  try {
-    request = raw.trim() ? (JSON.parse(raw) as AgentRunRequest) : {};
-  } catch (err) {
-    const failure: AgentRunResult = { ok: false, error: `Invalid JSON on stdin: ${String(err)}` };
-    if (stream) writeRecord({ kind: "result", result: failure });
-    else emitResult(failure);
-    process.exit(1);
-  }
-
-  if (!stream) {
-    try {
-      const result = await runAgent(request);
-      emitResult(result);
-      process.exit(result.ok ? 0 : 1);
-    } catch (err) {
-      emitResult({
-        ok: false,
-        error: err instanceof Error ? err.stack ?? err.message : String(err),
-      });
-      process.exit(1);
-    }
-    return;
-  }
-
-  const emit: EmitEvent = (event) => writeRecord({ kind: "event", event });
-  let result: AgentRunResult;
-  try {
-    result = await runAgent(request, emit);
-  } catch (err) {
-    result = { ok: false, error: err instanceof Error ? err.stack ?? err.message : String(err) };
-  }
-  // Streaming delivered the events live, so don't echo them in the terminal record.
-  writeRecord({ kind: "result", result: { ...result, events: [] } });
-  process.exit(result.ok ? 0 : 1);
+  const code = await runCli(raw, stream);
+  process.exit(code);
 }
 
-main();
+// Only run when this file is the process entry (`tsx src/cli.ts`); importing it is inert.
+if (isEntrypoint(import.meta.url)) {
+  void main();
+}
