+++ services/agent/src/server.ts
@@ -4,13 +4,21 @@
  * Same contract as the CLI, exposed over HTTP so the wrapper can run as its own
  * container (a sidecar) that the Python service calls in-network:
  *
- *   GET  /health -> { status: "ok" }
+ *   GET  /health -> runner identity ({ status, runner, protocol, engines, harnesses })
  *   POST /run    -> body is an AgentRunRequest, response is an AgentRunResult
  *
  * Uses Node's built-in http server (no framework dependency). Pi auth comes from
  * PI_CODING_AGENT_DIR / ~/.pi/agent, mounted into the container.
+ *
+ * `createAgentServer(run)` is the testable seam: it builds the server around an injectable
+ * engine runner so the HTTP behavior can be tested with a fake engine (no live harness).
  */
-import { createServer, type IncomingMessage, type ServerResponse } from "node:http";
+import {
+  createServer,
+  type IncomingMessage,
+  type Server,
+  type ServerResponse,
+} from "node:http";
 
 import type {
   AgentRunRequest,
@@ -20,6 +28,8 @@ import type {
 } from "./protocol.ts";
 import { runPi } from "./engines/pi.ts";
 import { runRivet } from "./engines/rivet.ts";
+import { runnerInfo } from "./version.ts";
+import { isEntrypoint } from "./entry.ts";
 
 const PORT = Number(process.env.PORT ?? 8765);
 
@@ -29,18 +39,21 @@ const PORT = Number(process.env.PORT ?? 8765);
 // request shape (a rivet request carries `harness`/`sandbox`).
 const DEFAULT_BACKEND = (process.env.AGENT_BACKEND ?? "auto").toLowerCase();
 
-function runAgent(
+/** Run one request through an engine. Tests inject a fake to avoid a live harness. */
+export type RunAgent = (
   request: AgentRunRequest,
   emit?: EmitEvent,
   signal?: AbortSignal,
-): Promise<AgentRunResult> {
+) => Promise<AgentRunResult>;
+
+const runAgent: RunAgent = (request, emit, signal) => {
   const backend = (request.backend ?? DEFAULT_BACKEND).toLowerCase();
   if (backend === "rivet") return runRivet(request, emit, signal);
   if (backend === "pi") return runPi(request, emit);
   return request.harness || request.sandbox
     ? runRivet(request, emit, signal)
     : runPi(request, emit);
-}
+};
 
 /**
  * Stream a run as NDJSON: one `{kind:"event"}` line per event the moment it is built, then
@@ -48,9 +61,10 @@ function runAgent(
  * with `Accept: application/x-ndjson`; the one-shot `/run` path is left untouched.
  */
 async function runAndStream(
-  req: IncomingMessage,
+  _req: IncomingMessage,
   res: ServerResponse,
   request: AgentRunRequest,
+  run: RunAgent,
 ): Promise<void> {
   res.writeHead(200, {
     "content-type": "application/x-ndjson",
@@ -75,7 +89,7 @@ async function runAndStream(
 
   let result: AgentRunResult;
   try {
-    result = await runAgent(request, emit, controller.signal);
+    result = await run(request, emit, controller.signal);
   } catch (err) {
     const message = err instanceof Error ? err.stack ?? err.message : String(err);
     result = { ok: false, error: message };
@@ -102,54 +116,68 @@ async function readBody(req: IncomingMessage): Promise<string> {
   return Buffer.concat(chunks).toString("utf8");
 }
 
-const server = createServer(async (req, res) => {
-  try {
-    if (req.method === "GET" && req.url === "/health") {
-      return send(res, 200, { status: "ok" });
-    }
-
-    if (req.method === "POST" && req.url === "/run") {
-      const raw = await readBody(req);
-      let request: AgentRunRequest;
-      try {
-        request = raw.trim() ? (JSON.parse(raw) as AgentRunRequest) : {};
-      } catch (err) {
-        return send(res, 400, { ok: false, error: `Invalid JSON: ${String(err)}` });
+/** Build the HTTP request listener around a given engine runner (the testable seam). */
+export function createRequestListener(
+  run: RunAgent,
+): (req: IncomingMessage, res: ServerResponse) => Promise<void> {
+  return async (req, res) => {
+    try {
+      if (req.method === "GET" && req.url === "/health") {
+        return send(res, 200, runnerInfo());
       }
 
-      const wantsStream = (req.headers["accept"] ?? "").includes(
-        "application/x-ndjson",
-      );
-      if (wantsStream) {
-        await runAndStream(req, res, request);
-        return;
+      if (req.method === "POST" && req.url === "/run") {
+        const raw = await readBody(req);
+        let request: AgentRunRequest;
+        try {
+          request = raw.trim() ? (JSON.parse(raw) as AgentRunRequest) : {};
+        } catch (err) {
+          return send(res, 400, { ok: false, error: `Invalid JSON: ${String(err)}` });
+        }
+
+        const wantsStream = (req.headers["accept"] ?? "").includes(
+          "application/x-ndjson",
+        );
+        if (wantsStream) {
+          await runAndStream(req, res, request, run);
+          return;
+        }
+
+        const result = await run(request);
+        return send(res, result.ok ? 200 : 500, result);
       }
 
-      const result = await runAgent(request);
-      return send(res, result.ok ? 200 : 500, result);
+      return send(res, 404, { ok: false, error: "Not found" });
+    } catch (err) {
+      const message = err instanceof Error ? err.stack ?? err.message : String(err);
+      return send(res, 500, { ok: false, error: message });
     }
+  };
+}
 
-    return send(res, 404, { ok: false, error: "Not found" });
-  } catch (err) {
-    const message = err instanceof Error ? err.stack ?? err.message : String(err);
-    return send(res, 500, { ok: false, error: message });
-  }
-});
+/** Create the sidecar HTTP server. Defaults to the real engine dispatch; tests pass a fake. */
+export function createAgentServer(run: RunAgent = runAgent): Server {
+  return createServer(createRequestListener(run));
+}
 
-// The rivet SDK can reject a background promise (e.g. an adapter install or the Daytona
-// preview SSE failing) outside any awaited path. Node's default turns that into an
-// uncaught exception that kills the whole process — taking every in-flight request with
-// it (the caller sees "Server disconnected"). Log and keep serving instead; the failing
-// run still returns its own error to its caller.
-process.on("unhandledRejection", (reason) => {
-  process.stderr.write(
-    `[pi-wrapper] unhandledRejection: ${reason instanceof Error ? (reason.stack ?? reason.message) : String(reason)}\n`,
-  );
-});
-process.on("uncaughtException", (err) => {
-  process.stderr.write(`[pi-wrapper] uncaughtException: ${err.stack ?? err.message}\n`);
-});
+// Only run as a server when this file is the process entry (`tsx src/server.ts`); importing
+// it (e.g. from a test) is inert.
+if (isEntrypoint(import.meta.url)) {
+  // The rivet SDK can reject a background promise (e.g. an adapter install or the Daytona
+  // preview SSE failing) outside any awaited path. Node's default turns that into an
+  // uncaught exception that kills the whole process — taking every in-flight request with
+  // it (the caller sees "Server disconnected"). Log and keep serving instead; the failing
+  // run still returns its own error to its caller.
+  process.on("unhandledRejection", (reason) => {
+    process.stderr.write(
+      `[pi-wrapper] unhandledRejection: ${reason instanceof Error ? (reason.stack ?? reason.message) : String(reason)}\n`,
+    );
+  });
+  process.on("uncaughtException", (err) => {
+    process.stderr.write(`[pi-wrapper] uncaughtException: ${err.stack ?? err.message}\n`);
+  });
 
-server.listen(PORT, () => {
-  process.stderr.write(`[pi-wrapper] http server listening on :${PORT}\n`);
-});
+  createAgentServer().listen(PORT, () => {
+    process.stderr.write(`[pi-wrapper] http server listening on :${PORT}\n`);
+  });
+}
